fix(clippy): resolve 286 test clippy warnings in frankenterm-core (ft-34l5u)
Eliminate all clippy warnings in frankenterm-core test code (--tests):
- float_cmp: add #[allow] to 8 test modules with intentional exact float comparisons
- needless_collect: replace collect().len() with count(), add allows where collect is needed
- naive_bytecount: allow in proptest_wavelet_tree (bytecount not a dependency)
- needless_continue: allow in proptest_work_stealing_deque
- needless_range_loop: allow in 5 proptest files with index-based test assertions
- overly_complex_bool_expr: allow tautological assertions in degradation/watchdog tests
- comparison_chain: allow intentional if-chains in proptest files
- match_wildcard_for_single_variants: allow in wait.rs tests
- similar_names: allow host_a_picks/host_b_picks in resize tests
- dead_code: allow unused helper strategies in proptest files
- approximate PI: replace 3.14 with std::f64::consts::PI
- Various auto-fixes via cargo clippy --fix (cloned→copied, useless vec!, etc.)

Combined with prior ft-159qj lib fixes, frankenterm-core now has 0 clippy warnings
for both --lib and --tests.

+++ crates/frankenterm-core/src/resize_scheduler.rs
@@ -3784,7 +3784,7 @@ mod tests {
                 1,
                 ResizeWorkClass::Interactive,
                 4,
-                100 + i as u64,
+                100 + i,
                 ResizeDomain::Ssh {
                     host: "remote1".into(),
                 },
